datecalc & timer built with webpack
