--- README.md
@@ -1,1 +1,0 @@
-# JS_bundle
