+++ hometask_2/index.html
@@ -2,8 +2,6 @@
 <html>
 <head>
 	<title>Datecalc & Timer</title>
-	<script src="https://cdnjs.cloudflare.com/ajax/libs/howler/2.2.1/howler.min.js" integrity="sha512-L6Z/YtIPQ7eU3BProP34WGU5yIRk7tNHk7vaC2dB1Vy1atz6wl9mCkTPPZ2Rn1qPr+vY2mZ9odZLdGYuaBk7dQ==" crossorigin="anonymous"></script>
-	<script type="module" src="src/main.js"></script>
 </head>
 <body>
 	<div class="control-btn">
