+++ hometask_2/src/common.js
