+++ hometask_2/src/main.js
